+++ src/main/java/com/practice/mongodb/config/ApplicationConfiguration.java
@@ -1,4 +1,4 @@
-package com.practice.mongodb;
+package com.practice.mongodb.config;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.springframework.beans.factory.annotation.Value;
