+++ src/main/java/com/practice/mongodb/document/Forecast.java
@@ -21,6 +21,8 @@ public class Forecast {
 
     private List<Period> periods;
 
+    public Forecast() {}
+
     public Forecast(UUID uuid, List<Period> periods) {
         this.uuid = uuid;
         this.periods = periods;
